requirements.txt
admin==0.0.1
asgiref==3.2.10
attrs==20.3.0
bcrypt==3.2.0
beautifultable==0.8.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
distro==1.5.0
Django==3.1.2
django-admin-thumbnails==0.2.5
django-adminlte3==0.1.6
django-braces==1.14.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-js-asset==1.2.2
django-model-utils==4.1.0
django-multiselectfield==0.1.12
django-online-users==0.3
django-star-ratings==0.9.1
django-template-tags==1.1.3
docker==4.4.3
docker-compose==1.28.4
dockerpty==0.4.1
docopt==0.6.2
fandogh-cli==1.24.1
gunicorn==20.0.4
idna==2.10
jsonschema==3.2.0
paramiko==2.7.2
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyrsistent==0.17.3
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
texttable==1.6.3
tzlocal==2.1
urllib3==1.26.2
websocket-client==0.57.0
zeep==4.0.0
